Rename cmd to installer
