+++ cmd/installer/subcommands/purge/command.go
@@ -7,7 +7,7 @@
 package purge
 
 import (
-	"github.com/DataDog/datadog-agent/cmd/updater/command"
+	"github.com/DataDog/datadog-agent/cmd/installer/command"
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/core/log/logimpl"
 	"github.com/DataDog/datadog-agent/pkg/installer"
@@ -32,7 +32,7 @@ func Commands(_ *command.GlobalParams) []*cobra.Command {
 func purgeFxWrapper() error {
 	return fxutil.OneShot(purge,
 		fx.Supply(core.BundleParams{
-			LogParams: logimpl.ForOneShot("UPDATER", "info", true),
+			LogParams: logimpl.ForOneShot("INSTALLER", "info", true),
 		}),
 		core.Bundle(),
 	)
