+++ cmd/installer/subcommands/run/command_test.go
@@ -3,19 +3,19 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package status
+package run
 
 import (
 	"testing"
 
-	"github.com/DataDog/datadog-agent/cmd/updater/command"
+	"github.com/DataDog/datadog-agent/cmd/installer/command"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
-func TestCommand(t *testing.T) {
+func TestRunCheckCmdCommand(t *testing.T) {
 	fxutil.TestOneShotSubcommand(t,
 		Commands(&command.GlobalParams{}),
-		[]string{"status"},
-		status,
+		[]string{"run"},
+		run,
 		func() {})
 }
