+++ cmd/installer/subcommands/status/status.tmpl
@@ -1,4 +1,4 @@
-Datadog Updater v{{ .Version }}
+Datadog Installer v{{ .Version }}
 {{ range $name, $package := .Packages }}
 {{ boldText $name }}
   State: {{ if $package.Stable }}{{ yellowText "unknown (unimplemented)" }}{{ else }} {{ redText "no stable version" }}{{ end }}
