+++ cmd/installer/user_windows.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023-present Datadog, Inc.
 
-// Package main implements 'updater'.
+// Package main implements 'installer'.
 package main
 
 func rootToDDInstaller() {}
